+++ app/admin/scrape/page.tsx
@@ -13,7 +13,7 @@ import { Button } from "@/components/button";
 import { PlusIcon, SparklesIcon } from "@heroicons/react/16/solid";
 import { TextLink } from "@/components/text";
 import { Badge } from "@/components/badge";
-import { product_links_from_crutchfield } from "@/lib/links";
+import { product_links_from_crutchfield } from "@/data/links";
 import "ldrs/lineSpinner"; // Default values shown
 
 import type { ScrapeLink } from "@/app/api/scrape/route";
